Update facts, others
