+++ test/integration/targets/eos_facts/tests/cli/invalid_subset.yaml
@@ -38,7 +38,7 @@
       # It's a failure
       - "result.failed == true"
       # Sensible Failure message
-      #- "result.msg == 'Bad subset'"
+      - "'Subset must be one of' in result.msg"
   ignore_errors: true
